the goal is to build a face recognition system, which includes building a face detector to locate the position of a face in an imageand a face identification model to recognize whose face it is by matching it to the existing database of faces.
In this project, we will build a face identification model by using an existing trained model.  The CNN model is taken from the Keras-OpenFace project. The architecture details aren't too important here, it's only useful to know that there is a fully connected layer with 128 hidden units followed by an L2 normalization layer on top of the convolutional base. These two top layers are referred to like the embedding layer from which the 128-dimensional embedding vectors can be obtained. Model training aims to learn an embedding f(x) of image x such that the squared L2 distance between all faces of the same identity is small and the distance between a pair of faces from different identities is large. This can be achieved with a triplet loss L that is minimized when the distance between an anchor image and a positive image (same identity) in embedding space is smaller than the distance between that anchor image and a negative image (different identity) by at least a margin. To demonstrate face recognition on a custom dataset, a small dataset is used. It consists of around 15-25 face images of 10 different persons. The metadata for each image (file and identity name) are loaded into memory for later processing. Test the Face Recognition system (the two models together) on a different dataset