The dataset is taken from Fake News Challenge, where the goal of the challenge is not to directly identify whether a headline or article is “fake” or not, which is arguably a highly subjective question, and one that even skilled humans may have difficulty answering. The challenge is organized around the more well-defined problem of “stance detection,” which involves comparing a headline with a body of text from a news article to determine what relationship (if any) exists between the two.
There are 4possible classifications:  1.The article text agrees with the headline.  2.The article text disagrees with the headline.  3.The article text is a discussion of the headline, without taking a position on it.4.The article text is unrelated to the headline (i.e. it doesn’t address the same topic).
Steps involved :
Step 1: Understanding how to convert words into vectors and make them ready for the model. This is an important step and will be the same for almost any problem in NLP when dealing with words. Step 2: Understanding the working of Sequential Models like RNN, LSTM.. and practically building them in Keras and build the model ready. Step 3: Train the model and report the accuracy score. Step 4: Add checkpoints as callbacks to the model and save the model for each epoch. This callback saves the model after each epoch, which can be handy for when you are running long-term training. By loading the saved weight files we can start training the model which starts the learning from there.